+++ src/acconeer/exptool/a121/_core/entities/__init__.py
@@ -18,6 +18,7 @@ from .containers import (
     Metadata,
     PersistentRecord,
     Record,
+    RecordException,
     Result,
     ResultContext,
     SensorCalibration,