+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Fixed
 - A121 recorder crashing when stopping due to empty chunk buffer
+- A121 recorder can't load recording without calibration
 
 ## v5.4.0